refactor(hub-common): use truncateSlug and getSlugMaxLength
